feat(wallet): enhance wallet functionality with new account management and transaction features
- Added account resolution and setting endpoints for bank details.
- Implemented withdrawal functionality with fee calculations.
- Introduced transaction history retrieval for wallet transactions.
- Updated wallet balance fetching and wallet confirmation logic.
- Enhanced fund wallet process with payment method handling.
- Created new request classes for validation of account details and fund requests.
- Added property settings management for landlords and tenants.
- Introduced rent payment and installment management features.
- Updated database migrations for new fields and tables related to wallet and rent payments.
- Added new views for payment verification.

+++ database/migrations/2024_04_25_212949_create_wallets_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(User::class, 'user_id');
             $table->string('uuid');
-            $table->double('balance')->default(0);
-            $table->double('total_credit')->default(0);
-            $table->double('total_debit')->default(0);
+            $table->decimal('balance', 20, 2)->default(0);
+            $table->decimal('total_credit', 20, 2)->default(0);
+            $table->decimal('total_debit', 20, 2)->default(0);
             $table->string('bank')->nullable();
             $table->string('bank_code')->nullable();
             $table->string('account_number')->nullable();
